pyproject.toml
[project]
name = "rosetta"
version = "0.1.0"
description = "Multi-agent drug discovery system powered by Claude Agent SDK"
requires-python = ">=3.12"
dependencies = [
    # Agent SDK
    "claude-agent-sdk>=0.1.44",
    # Data models & serialization
    "pydantic>=2.0",
    "pyyaml>=6.0",
    # Core data science (for analysis modules)
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "statsmodels>=0.14",
    # API access
    "requests>=2.31",
    # Utilities
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
bio = [
    # Heavier bioinformatics deps (run in Docker or install separately)
    "scanpy>=1.9",
    "anndata>=0.9",
    "pydeseq2>=0.4",
    "squidpy>=1.4",
    "pyliftover>=0.4",
    "gseapy>=1.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
